cksum {bitops}	R Documentation

Compute Check Sum

Description

return a cyclic redundancy checksum for each element in the argument.

Usage

cksum(a)

Arguments

a 	 coerced to character vector

Details

NA's appearing in the argument are returned as NA's.

The default calculation is identical to that given in pseudo-code in the ACM article (in the References).

Value

numeric vector of length a.

Author(s)

Steve Dutky sdutky@terpalum.umd.edu

References

Fashioned from cksum(1) UNIX command line utility, i.e., man cksum.

Dilip V. Sarwate (1988). Computation of Cyclic Redundancy Checks Via Table Lookup, Communications of the ACM, August 1988. vol 31, No.8 page 1008-1013

Examples

   b <- "I would rather have a bottle in front of me than frontal lobotomy\n"
   cksum(b) == 1342168430 ## -> TRUE

[Package bitops version 1.0-4.1 Index]